MAR-848 Removed the getRoutesList helper from nuxt.config
